Spanish translation
Hello! I wanted to contribute my grain of sand! 😊
It's my first time here on github so I'm not really sure if the code or PR turned out as it should, I copied the base translation from the translate.py file and started translating from there. I apologize if there are problems. 

I will be watching in case there are more changes to translate in the future.