linux/pyinstaller: exclude unneeded libraries
Excluding these Linux libraries doesn't seem to cause any side-effects
at runtime (they're also not present in the AppImage).

And the benefit of this, is a reduction of 15 MiB in the final size of
the Linux binary, which should also slightly reduce the app's startup
time (but don't quote me on that). :)